[add] 設定ファイルの項目から build.input を削除

+++ src/command/utils/build.rs
@@ -1,47 +1,13 @@
 use std::fs;
 use std::process::Command as StdCommand;
 
-use crate::config::build::InputType;
 use crate::config::Config;
 
 pub fn build(config: &Config) -> anyhow::Result<()> {
-    // 1. src/main.asm or src/main.sb をチェック
-    match &config.build.input {
-        InputType::Asm => if !fs::exists("./src/main.asm")? {
-            return Err(anyhow::anyhow!("src/main.asm not found."));
-        }
-        InputType::Sblang => if !fs::exists("./src/main.sb")? {
-            return Err(anyhow::anyhow!("src/main.sb not found."));
-        }
+    // 1. 準備
+    if !fs::exists("./src/main.sb")? {
+        return Err(anyhow::anyhow!("src/main.sb not found."));
     }
-
-    // 2. ビルド
-    match &config.build.input {
-        InputType::Asm => build_asm(config),
-        InputType::Sblang => build_sblang(config),
-    }
-}
-
-fn build_asm(_: &Config) -> anyhow::Result<()> {
-    // 1. target 準備
-    let _ = fs::remove_dir("target/out");
-    fs::create_dir_all("target/out/hex")?;
-
-    // 2. アセンブル
-    let status = StdCommand::new("sb-assembler")
-        .arg("./src/main.asm")
-        .arg("./target/out/hex/data.hex")
-        .arg("./target/out/hex/inst.hex")
-        .status()?;
-    if !status.success() {
-        return Err(anyhow::anyhow!("Assemble process failed."));
-    }
-
-    Ok(())
-}
-
-fn build_sblang(config: &Config) -> anyhow::Result<()> {
-    // 1. target 準備
     let _ = fs::remove_dir("target/build");
     let _ = fs::remove_dir("target/out");
     fs::create_dir_all("target/build")?;
